UPDATE rol y usuarios

+++ build/controller/controllerUsuario.js
@@ -8,24 +8,26 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
+};
 Object.defineProperty(exports, "__esModule", { value: true });
-const usuarios = require('./../../models').Usuarios;
+const globales_1 = __importDefault(require("./../utils/globales"));
+const rols = require("./../../models").rols;
+const seguridad_2 = require("./../utils/seguridad");
+const seguridad_1 = require("./../utils/seguridad");
+const usuarios = require("./../../models").Usuarios;
 class usuariosController {
-    getData(req, res) {
+    addUsuario(req, res) {
         return __awaiter(this, void 0, void 0, function* () {
-            usuarios.findAll().then((data) => {
-                res.status(200).json(data);
+            let token = true;
+            if (!token) {
+                res.status(401).json({ log: "Token invalido." });
                 return;
-            }, (err) => {
-                res.status(500).json({ log: "Error!! No hay datos en la base" });
-                console.log(err);
-                return;
-            });
-        });
-    }
-    postData(req, res) {
-        return __awaiter(this, void 0, void 0, function* () {
-            let usuario = {
+            }
+            let { hash } = req.body;
+            //aqui desencriptar los datos
+            let data = {
                 cedula: req.body.cedula,
                 nombre: req.body.nombre,
                 apellido: req.body.apellido,
@@ -34,14 +36,28 @@ class usuariosController {
                 direccion: req.body.direccion,
                 contrasenia: req.body.contrasenia,
                 rol: req.body.rol,
-                createdAt: new Date(),
-                updatedAt: new Date()
             };
-            usuarios.create(usuario).then((rs) => {
-                res.status(200).json(rs);
+            let hashInterno = seguridad_2.Seguridad.hashJSON(data);
+            //aqui se debe desencriptar el hash
+            data.createdAt = new Date();
+            if (hashInterno != hash) {
+                res
+                    .status(401)
+                    .json({ log: "Violación de integridad de datos, hash invalido." });
+                return;
+            }
+            usuarios.create(data).then((resp) => {
+                if (resp._options.isNewRecord) {
+                    res.status(202).json({
+                        log: "Ingresado",
+                        uri: globales_1.default.globales.urlBaseUsuario + resp.dataValues.cedula,
+                    });
+                    return;
+                }
+                res.status(200).json({ log: "No se ingresaron los datos." });
                 return;
             }, (err) => {
-                res.status(500).json({ log: "Error!! No se pudo crear el usuario" });
+                res.status(500).json({ log: "Error" });
                 console.log(err);
                 return;
             });
@@ -56,16 +72,30 @@ class usuariosController {
             }
             id = Number(id);
             if (Number.isInteger(id) == false) {
-                res.status(500).json({ log: "El ID introducido no es valido, debe ser un entero." });
+                res
+                    .status(500)
+                    .json({ log: "El ID introducido no es valido, debe ser un entero." });
                 return;
             }
-            usuarios.findOne({
+            usuarios
+                .findOne({
                 where: {
-                    cedula: id
-                }
-            }).then((data) => {
+                    cedula: id,
+                },
+                attributes: ["cedula", "nombre", "apellido", "direccion", "rol"],
+                include: [
+                    {
+                        model: rols,
+                        required: true,
+                        attributes: ["idrol"],
+                    },
+                ],
+            })
+                .then((data) => {
                 if (data == null) {
-                    res.status(404).json({ log: "No Existen datos a mostrar para el ID." });
+                    res
+                        .status(404)
+                        .json({ log: "No Existen datos a mostrar para el ID." });
                     return;
                 }
                 res.status(200).json(data);
@@ -76,53 +106,31 @@ class usuariosController {
             });
         });
     }
-    deleteData(req, res) {
+    deleteUsuario(req, res) {
         return __awaiter(this, void 0, void 0, function* () {
-            let { id } = req.params;
+            let id = req.params.id;
+            if (isNaN(id)) {
+                res.status(500).json({ log: "El ID introducido no es valido." });
+                return;
+            }
+            id = Number(id);
+            if (Number.isInteger(id) == false) {
+                res
+                    .status(500)
+                    .json({ log: "El ID introducido no es valido, debe ser un entero." });
+                return;
+            }
             usuarios.destroy({ where: { cedula: id } }).then((data) => {
                 if (data == 1) {
-                    res.status(200).json({ log: "Usuario eliminado exitosamente" });
+                    res.status(200).json({ log: "Eliminado" });
                     return;
                 }
                 else {
-                    res.status(200).json({ log: "No existe el usuario" });
+                    res.status(200).json({ log: "Sin datos a eliminar." });
                     return;
                 }
             }, (err) => {
-                res.status(500).json({ log: "Error!!" });
-                console.log(err);
-                return;
-            });
-        });
-    }
-    updateData(req, res) {
-        return __awaiter(this, void 0, void 0, function* () {
-            let { id } = req.params;
-            console.log(id);
-            let { cedula, nombre, apellido, telefono, email, direccion, contrasenia, rol, createdAt } = req.body;
-            let est = {
-                cedula: cedula,
-                nombre: nombre,
-                apellido: apellido,
-                telefono: telefono,
-                email: email,
-                direccion: direccion,
-                contrasenia: contrasenia,
-                rol: rol,
-                createdAt: createdAt,
-                updatedAt: new Date()
-            };
-            usuarios.update(est, { where: { cedula: id } }).then((rs) => {
-                if (rs[0] === 1) {
-                    res.status(200).json({ log: "El usuario se actualizó." });
-                    return;
-                }
-                else {
-                    res.status(200).json({ log: "No se pudo modificar el usuario" });
-                    return;
-                }
-            }, (err) => {
-                res.status(500).json({ log: "Error del servidor" });
+                res.status(500).json({ log: "Error" });
                 console.log(err);
                 return;
             });
